+++ lib/ahoy_email.rb
@@ -21,7 +21,8 @@ module AhoyEmail
     utm_term: nil,
     utm_content: nil,
     utm_campaign: proc {|message, mailer| mailer.action_name },
-    user: proc{|message, mailer| (message.to.size == 1 ? User.where(email: message.to.first).first : nil) rescue nil }
+    user: proc{|message, mailer| (message.to.size == 1 ? User.where(email: message.to.first).first : nil) rescue nil },
+    mailer: proc{|message, mailer| "#{mailer.class.name}##{mailer.action_name}" }
   }
 
   self.subscribers = []