+++ lib/ahoy_email/processor.rb
@@ -19,9 +19,10 @@ module AhoyEmail
         track_open if options[:open]
         track_links if options[:utm_params] or options[:click]
 
-        # save
+        ahoy_message.mailer = options[:mailer] if ahoy_message.respond_to?(:mailer=)
         ahoy_message.subject = message.subject if ahoy_message.respond_to?(:subject=)
         ahoy_message.content = message.to_s if ahoy_message.respond_to?(:content=)
+
         ahoy_message.save
         message["Ahoy-Message-Id"] = ahoy_message.id
       end