+++ lib/generators/ahoy_email/templates/install.rb
@@ -10,6 +10,7 @@ class <%= migration_class_name %> < ActiveRecord::Migration
 
       # optional
       # feel free to remove
+      t.string :mailer
       t.text :subject
       t.text :content